requirements.txt
fyers-apiv3
pyotp
requests
pytz
schedule
python-dotenv
setuptools
twilio
